class: Switch Chinese font to 'Noto Serif CJK' since it's more common
Actually, the 'Noto Serif CJK' font and the previous 'Source Han Serif'
font are the same font but with different names.  However, the former is
more common and ready to use; for example, it's generally installed by
default on many Linux distros and it's available on Overleaf.

Update the README accordingly.